requirements.txt
# Shared Dependencies for Medical Imaging Backend Services

# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==21.2.0
pydantic==2.6.1
pydantic-settings==2.1.0

# Database - PostgreSQL
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1

# Database - Redis
redis==5.0.1
aioredis==2.0.1

# Message Queue
aio-pika==9.3.1
celery==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-keycloak==3.9.0
python-multipart==0.0.9
cryptography==41.0.7  # For encryption

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Validation
email-validator==2.1.0.post1

# Utilities
python-dotenv==1.0.1
pytz==2024.1

# Logging & Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.5.0

# Medical Imaging (for services that need it)
pydicom==2.4.4
SimpleITK==2.4.1
nibabel==5.2.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
